Add GameOptions.java as component to main menu ui
